BUGFIX. Scrollbar customization using CSS was ineffective
Error:
- The previous implementation of customizing the scrollbar using CSS did not work as expected.

Solution:
- To achieve consistent and cross-browser `scrollbar` customization, the decision was made to utilize the `perfect-scrollbar`.
`Perfect Scrollbar` is a JavaScript library that dynamically generates HTML elements to create custom scrollbars and applies styles and sizes to ensure visual consistency and responsiveness.
The library dynamically calculates and applies appropriate sizes and positions to these elements based on the content and dimensions of the container.
It ensures that scrollbar components are visually consistent and responsive to user interactions, such as scrolling or resizing the container. Allows easy customization of scrollbars without relying on the browser's default styles.

BREAKING CHANGE:
- The `perfect-scrollbar` package must be installed.

References:
- https://perfectscrollbar.com/
- https://www.npmjs.com/package/perfect-scrollbar
- https://mdbootstrap.com/snippets/standard/marveluck/3498209?utm_source=GitHub&utm_medium=PerfectScrollbar#js-tab-view