BUGFIX. Get an error when compiling js with the `gulp` command
Error:
- Starting 'jsCompile'...
'jsCompile' errored after 341 ms
Error in plugin "gulp-babel"
Message:
    /Applications/MAMP/htdocs/calendar/src/js/pages/_pages-calendar-1.js: Couldn't find intersection

Solution:
- Update the `babel` configuration in the `jsCompile` gulp function.
Call the `babel-preset-minify` package with the name `minify` instead and disable the `builtIns` and `mangle` options.

Reason:
- The reason behind this change may be the use of a newer version of Babel or a related library. The syntax `babel-preset-minify` has been replaced by simply `minify with additional options to customize the behavior of the minification process.
The `"builtIns": false` and `"mangle": false` options are disabled in the Babel configuration to prevent the transformation of new native JavaScript methods and objects and to keep variable and property names non-obfuscated, respectively.

References:
- https://github.com/babel/minify/issues/904#issuecomment-1525822017
- https://www.npmjs.com/package/babel-preset-minify