BUGFIX. Calendar does not display on a single sheet when printing from Safari
Expected:
- When generating a PDF, the calendar should be displayed on a single sheet:
In `landscape` orientation` it should have a `4x3` layout.
In `portrait` orientation` it should have a `3x4` layout.

Error:
- Safari browser exhibits incompatibilities with print-related CSS properties (tested on mac with Safari, and iphone with Safari and Chrome):
Not working `@media print and (orientation: portrait)` and `@media print and (orientation: landscape)` media queries.
Can not remove the `margin` of the page although use the css property `margin: 0cm !important` in the `@page` selector.
Can not force the `landscape` of he page although use the css property `size: A4 landscape` in the `@page` selector.

Solution:
- Implement a grid layout using `display: grid` and `grid-template-columns` instead of `display: flex`.
- Convert the units from `rem` and `px` to `in`, `cm`, `mm` and `pt` with the `convertUnit` sass function.
- Optimize sizes by using `1fr` for flexible column sizing instead of fixed sizes.
- Reduce the sizes for the `font-size`, `padding`, `margin`, `width`, `height`...
- Removes the extra space created by the default value of the `border-spacing` css property from the `calendar__table` class.
- Remove `@media print and (orientation: portrait)` and `@media print and (orientation: landscape)` media queries.
- Set a minimun `margin` of 1cm for all browsers.