GULP. Create `jsTest` task to run and debug `js`.
Reason:
- Breakpoints are crucial for accurate identification and resolution of code problems, especially in complex applications.
Gulp concatenates and minifies the individual JavaScript files, and the file linked in the HTML is the one that Visual Studio Code (VSC) looks at during debugging. Therefore, debugging directly from the individual files is not possible due to the compilation process.
The concatenated and minified `scripts.min.js` file produced by Gulp for production use, cannot be easily debugged.

Solution:
- The `jsTest` task generates an unminified JavaScript file, `scripts.js`, that retains the format and structure of the original code, which improves code readability and simplifies the debugging process, as it closely resembles the individual files before concatenation and minification.
- You can now set breakpoints and debug directly from the VSC editor using the `scripts.js` file instead of `scripts.min.js` in the HTML file.
- A configuration file, `launch.json`, is generated from VSC updating the port to match the one used in the Gulp configuration (3000).
This allows VSC to monitor the execution of the files from the Chrome browser, stopping at the assigned breakpoints.
- Once errors are found, they can be corrected in the individual files that Gulp concatenates, and the HTML file needs to be updated to reference the production `scripts.min.js` file again.