GULP. Removing comments in the css file in production
Steps:
- Install the `gulp-stript-css-comments` package.
- Import the `gulp-stript-css-comments` package as `gulpStripCssComments` in the `gulpfile.js` file.
- Use the `gulpStripCssComments` function in the `sassCompile` function to remove css comments from the `dist/css/styles.min.css` file.

Reason:
- Comments in CSS files are useful during development to provide clarifications, annotations or instructions for developers. However, once the project is ready to be released to production, these comments are no longer needed and can take up unnecessary space in the final CSS file.
By removing comments in the production version, we can reduce the size of the CSS file, which can have a positive impact on page load performance, as less data will be sent to the user's browser. In addition, a cleaner CSS file without comments can make the code easier to maintain and more readable.

References:
- https://www.npmjs.com/package/gulp-strip-css-comments