REFACTOR. Use the `transform` css property instead `calc` to align the element
Reason:
- When using `calc` to calculate the position coordinates (`top` and `left`), complex mathematical calculations must be performed at render time. This can generate additional load for the browser and affect page performance, especially when frequent layout changes are applied.

Steps:
- The advantage of using the `transform` CSS property instead of `calc` to align the element is that `transform` provides a more efficient and accurate way of centering elements.
Using `transform: translate(-50%, -50%)`, the same result of vertical and horizontal centering is achieved, but in a simpler and more optimized way. The `translate` property allows you to move the element relative to its own size, so that `-50%` translates as half the height and half the width of the element. This ensures that the element is perfectly centered, without the need for complex calculations.