BUGFIX. Add scrollbar snapping functionality
Error:
- To resolve the issue that prevented dragging the months to scroll through the calendar, I have removed the CSS properties `scroll-snap-type: x mandatory` and `scroll-snap-align: center`.
The CSS property `scroll-snap-type` was used in the container to control the direction of horizontal scrolling and determine whether content items should snap to the edges of their container. On the other hand, the CSS property `scroll-snap-align` was used on each element inside the container, to specify that it will align centered with respect to its container.

Steps:
- To replace the removed CSS properties, I inserted the `panelsnap` package into the project.
Install the `panelsnap` package with the `npm install panelsnap` command.
PanelSnap is a JavaScript agnostic library. This means that it works in all JavaScript projects, whether you use Vue, React, jQuery or just JavaScript. It can snap both horizontally and vertically, connect to menus and fire events based on user behavior.
- Import the `panelsnap` package as `PanelSnap` in the `_pages-calendar-1.js` file.
- Create the new `scrollbarSnap` function to configure and activate the snapping behavior.

References:
- https://alvarotrigo.com/blog/why-not-to-use-css-scroll-snap/
- https://panelsnap.com/
- https://panelsnap.com/demos/horizontal/
- https://github.com/guidobouman/panelsnap